feat: zero constraints stark (#1489)
