+++ circuits/src/rangecheck/stark.rs
@@ -1,65 +1,10 @@
-use std::marker::PhantomData;
+use super::columns::RangeCheckColumnsView;
+use crate::columns_view::NumberOfColumns;
+use crate::unstark::Unstark;
 
-use mozak_circuits_derive::StarkNameDisplay;
-use plonky2::field::extension::{Extendable, FieldExtension};
-use plonky2::field::packed::PackedField;
-use plonky2::hash::hash_types::RichField;
-use plonky2::iop::ext_target::ExtensionTarget;
-use plonky2::plonk::circuit_builder::CircuitBuilder;
-use starky::constraint_consumer::{ConstraintConsumer, RecursiveConstraintConsumer};
-use starky::evaluation_frame::StarkFrame;
-use starky::stark::Stark;
-
-use super::columns::{self, RangeCheckColumnsView};
-use crate::columns_view::HasNamedColumns;
-
-#[derive(Copy, Clone, Default, StarkNameDisplay)]
 #[allow(clippy::module_name_repetitions)]
-pub struct RangeCheckStark<F, const D: usize> {
-    pub _f: PhantomData<F>,
-}
-
-impl<F, const D: usize> HasNamedColumns for RangeCheckStark<F, D> {
-    type Columns = RangeCheckColumnsView<F>;
-}
-
-const COLUMNS: usize = columns::NUM_RC_COLS;
-const PUBLIC_INPUTS: usize = 0;
-
-impl<F: RichField + Extendable<D>, const D: usize> Stark<F, D> for RangeCheckStark<F, D> {
-    type EvaluationFrame<FE, P, const D2: usize> = StarkFrame<P, P::Scalar, COLUMNS, PUBLIC_INPUTS>
-
-    where
-        FE: FieldExtension<D2, BaseField = F>,
-        P: PackedField<Scalar = FE>;
-    type EvaluationFrameTarget =
-        StarkFrame<ExtensionTarget<D>, ExtensionTarget<D>, COLUMNS, PUBLIC_INPUTS>;
-
-    // NOTE: Actual range check happens in RangeCheckU8Stark. A CrossTableLookup
-    // between RangeCheckStark and others like MemoryStark and CpuStark ensure
-    // that both have same value. A CrossTableLookup between RangeCheckStark and
-    // RangeCheckU8Stark ensures that each limb from this stark are covered
-    // in RangeCheckU8Stark.
-    // Here we check if limbs are formed properly
-    fn eval_packed_generic<FE, P, const D2: usize>(
-        &self,
-        _vars: &Self::EvaluationFrame<FE, P, D2>,
-        _yield_constr: &mut ConstraintConsumer<P>,
-    ) where
-        FE: FieldExtension<D2, BaseField = F>,
-        P: PackedField<Scalar = FE>, {
-    }
-
-    fn eval_ext_circuit(
-        &self,
-        _builder: &mut CircuitBuilder<F, D>,
-        _vars: &Self::EvaluationFrameTarget,
-        _yield_constr: &mut RecursiveConstraintConsumer<F, D>,
-    ) {
-    }
-
-    fn constraint_degree(&self) -> usize { 3 }
-}
+pub type RangeCheckStark<F, const D: usize> =
+    Unstark<F, D, RangeCheckColumnsView<F>, { RangeCheckColumnsView::<()>::NUMBER_OF_COLUMNS }>;
 
 #[cfg(test)]
 mod tests {
