+++ circuits/src/register/zero_write/stark.rs
@@ -1,55 +1,7 @@
-use std::marker::PhantomData;
-
-use mozak_circuits_derive::StarkNameDisplay;
-use plonky2::field::extension::{Extendable, FieldExtension};
-use plonky2::field::packed::PackedField;
-use plonky2::hash::hash_types::RichField;
-use plonky2::iop::ext_target::ExtensionTarget;
-use plonky2::plonk::circuit_builder::CircuitBuilder;
-use starky::constraint_consumer::{ConstraintConsumer, RecursiveConstraintConsumer};
-use starky::evaluation_frame::StarkFrame;
-use starky::stark::Stark;
-
 use super::columns::RegisterZeroWrite;
-use crate::columns_view::{HasNamedColumns, NumberOfColumns};
+use crate::columns_view::NumberOfColumns;
+use crate::unstark::Unstark;
 
-#[derive(Clone, Copy, Default, StarkNameDisplay)]
 #[allow(clippy::module_name_repetitions)]
-pub struct RegisterZeroWriteStark<F, const D: usize>(PhantomData<F>);
-
-impl<F, const D: usize> HasNamedColumns for RegisterZeroWriteStark<F, D> {
-    type Columns = RegisterZeroWrite<F>;
-}
-
-const COLUMNS: usize = RegisterZeroWrite::<()>::NUMBER_OF_COLUMNS;
-const PUBLIC_INPUTS: usize = 0;
-
-impl<F: RichField + Extendable<D>, const D: usize> Stark<F, D> for RegisterZeroWriteStark<F, D> {
-    type EvaluationFrame<FE, P, const D2: usize> = StarkFrame<P, P::Scalar, COLUMNS, PUBLIC_INPUTS>
-
-    where
-        FE: FieldExtension<D2, BaseField = F>,
-        P: PackedField<Scalar = FE>;
-    type EvaluationFrameTarget =
-        StarkFrame<ExtensionTarget<D>, ExtensionTarget<D>, COLUMNS, PUBLIC_INPUTS>;
-
-    /// Constraints for the [`RegisterZeroWriteStark`]
-    fn eval_packed_generic<FE, P, const D2: usize>(
-        &self,
-        _vars: &Self::EvaluationFrame<FE, P, D2>,
-        _yield_constr: &mut ConstraintConsumer<P>,
-    ) where
-        FE: FieldExtension<D2, BaseField = F>,
-        P: PackedField<Scalar = FE>, {
-    }
-
-    fn eval_ext_circuit(
-        &self,
-        _builder: &mut CircuitBuilder<F, D>,
-        _vars: &Self::EvaluationFrameTarget,
-        _yield_constr: &mut RecursiveConstraintConsumer<F, D>,
-    ) {
-    }
-
-    fn constraint_degree(&self) -> usize { 3 }
-}
+pub type RegisterZeroWriteStark<F, const D: usize> =
+    Unstark<F, D, RegisterZeroWrite<F>, { RegisterZeroWrite::<()>::NUMBER_OF_COLUMNS }>;
